#13090: iOS 13: Selected image disappears the first time in Chat screen.
Problem: In iOS 13 when reloading the collection view a scrolling it to the right immediately when there is only 1 item cause the item to animate offscreen. This is caused by invoking reloadData and scrolltoItem methods one after other immediately .
Solution: Using G.C.D to add the the scroll to main queue will allow some time for collection view to reload.